requirements.txt
streamlit
fastapi
pydantic
uvicorn
langchain
langgraph
duckduckgo-search
google-generativeai
langchain-mcp-adapters
langsmith
duckduckgo_search 